move NestDictionaryEntries out of Bindings
